back_end/requirements.txt
"pymongo[srv]"
python-dotenv
certifi
browser-cookie3
requests